# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Receitas Culinarias",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/recipes": {
            "get": {
                "tags": [
                    "Recipes"
                ],
                "summary": "Listagem de receitas",
                "description": "Retorna todas as receitas cadastradas de forma paginada.",
                "operationId": "e6ebf01e62f938e7ce367c7b2fc94110",
                "responses": {
                    "200": {
                        "description": "Lista de receitas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Bolo de Cenoura"
                                            },
                                            "user_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "category_id": {
                                                "type": "integer",
                                                "example": 3
                                            },
                                            "preparation_time": {
                                                "type": "integer",
                                                "example": 60
                                            },
                                            "portion": {
                                                "type": "integer",
                                                "example": 8
                                            },
                                            "preparation_mode": {
                                                "type": "string",
                                                "example": "Misture todos os ingredientes e leve ao forno."
                                            },
                                            "ingredients": {
                                                "type": "string",
                                                "example": "3 ovos, 2 xícaras de farinha, 1 xícara de óleo"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2025-09-10T12:00:00Z"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2025-09-10T12:10:00Z"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Recipes"
                ],
                "summary": "Criar nova receita",
                "description": "Cadastra uma nova receita no sistema.",
                "operationId": "5ed3009a1c8ce5b784cac78533de2c73",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "user_id",
                                    "category_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Bolo de Chocolate"
                                    },
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "preparation_time": {
                                        "type": "integer",
                                        "example": 45
                                    },
                                    "portion": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "preparation_mode": {
                                        "type": "string",
                                        "example": "Bata os ovos, açúcar e chocolate, asse por 40 min."
                                    },
                                    "ingredients": {
                                        "type": "string",
                                        "example": "4 ovos, 2 xícaras de açúcar, 1 xícara de chocolate em pó"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Receita cadastrada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Receita cadastrada com sucesso."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação"
                    },
                    "500": {
                        "description": "Erro ao criar a receita"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/recipes/{id}": {
            "get": {
                "tags": [
                    "Recipes"
                ],
                "summary": "Exibir receita",
                "description": "Busca uma receita específica pelo ID.",
                "operationId": "e59df102512a161f3e7be88aaa85a0e1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da receita",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Receita encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Lasanha à Bolonhesa"
                                        },
                                        "user_id": {
                                            "type": "integer",
                                            "example": 2
                                        },
                                        "category_id": {
                                            "type": "integer",
                                            "example": 4
                                        },
                                        "preparation_time": {
                                            "type": "integer",
                                            "example": 90
                                        },
                                        "portion": {
                                            "type": "integer",
                                            "example": 6
                                        },
                                        "preparation_mode": {
                                            "type": "string",
                                            "example": "Monte camadas de massa, carne e queijo, leve ao forno."
                                        },
                                        "ingredients": {
                                            "type": "string",
                                            "example": "Massa de lasanha, molho bolonhesa, queijo mussarela"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-09-09T10:00:00Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-09-09T12:00:00Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Receita não encontrada"
                    },
                    "500": {
                        "description": "Erro ao buscar a receita"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Recipes"
                ],
                "summary": "Atualizar receita",
                "description": "Atualiza os dados de uma receita existente.",
                "operationId": "573ece0098b0f511b333f1406621b99c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da receita",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Bolo de Fubá"
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "preparation_time": {
                                        "type": "integer",
                                        "example": 50
                                    },
                                    "portion": {
                                        "type": "integer",
                                        "example": 12
                                    },
                                    "preparation_mode": {
                                        "type": "string",
                                        "example": "Misture os ingredientes secos e líquidos, asse por 50 minutos."
                                    },
                                    "ingredients": {
                                        "type": "string",
                                        "example": "Fubá, ovos, açúcar, leite"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Receita atualizada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Receita atualizada com sucesso."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Receita não encontrada"
                    },
                    "500": {
                        "description": "Erro ao atualizar a receita"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Recipes"
                ],
                "summary": "Deletar receita",
                "description": "Remove uma receita existente pelo ID.",
                "operationId": "a527903d2d7195e7456416e97f55faf2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da receita",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Receita deletada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Receita deletada com sucesso."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Receita não encontrada"
                    },
                    "500": {
                        "description": "Erro ao deletar a receita"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
...
}